re-organize dispatch files
